+++ operator/apis/http/v1alpha1/httpscaledobject_types.go
@@ -31,7 +31,9 @@ type ScaleTargetRef struct {
 	// The name of the service to route to
 	Service string `json:"service"`
 	// The port to route to
-	Port int32 `json:"port"`
+	Port int32 `json:"port,omitempty"`
+	// The port to route to referenced by name
+	PortName string `json:"portName,omitempty"`
 }
 
 // ReplicaStruct contains the minimum and maximum amount of replicas to have in the deployment
@@ -88,6 +90,8 @@ type HTTPScaledObjectSpec struct {
 	// +optional
 	PathPrefixes []string `json:"pathPrefixes,omitempty"`
 	// The name of the deployment to route HTTP requests to (and to autoscale).
+	// Including validation as a requirement to define either the PortName or the Port
+	// +kubebuilder:validation:XValidation:rule="has(self.portName) != has(self.port)",message="must define either the 'portName' or the 'port'"
 	ScaleTargetRef ScaleTargetRef `json:"scaleTargetRef"`
 	// (optional) Replica information
 	// +optional